Add chain selector
The chain selector for the classes "avatar" and "proportioned" has
been added followed by a declaration block containing values for
height and width. This selector has been added as the img elements
share a class name so the chain selector is important to distinguish
the images and give each one a unique style.