Move para and small para classes
The para class' declaration block has been moved above the small para
class. This will make sure that the small para class' declarations
will be applied to the second p element. Switching the position of
these declaration blocks fixes the cascade as in this case, the last
defined rule will be applied to the second paragraph element.