Move button and confirm classes and declarations
The button class and its corresponding declaration block have been
moved above the confirm class. This allows the confirm class'
declarations to be applied to the first button. Fixing the styling
applied in this case was also a matter of switching the classes around
so that the last defined rule would be applied.