Add margin and padding to ul and a
The margin and padding properties have been added to the ul elements
whereas the margin property has been added to the a elements. These
changes have been made in order to space out these elements so that
they are more easily accessible and able to be seen as separate
elements by the user.