Edit values for flex properties, add gap property
The values for justify-content in the header, margin in the ul and
padding in the ul declaration blocks have been edited in order to
align the content on the page in the specified manner. The gap
property has also been added to the ul element in order to ensure
that each list element has a visible gap between them, making the
links stand out from one another and making them more visually
pleasing.