Add padding, gap and vertically align header
The padding property has been added to the header class in order to
ensure that there is space between the edge of the header and the
content it contains.

The gap property has been added to ensure that there is a gap between
each flex item in the header flex container.

The align-items property has also been added to the header flex
container in order to ensure that each flex item is vertically centred.