Create flex container, add gap property
The display rule has been declared underneath the ul element selector
and has had its value set to flex. This will make the ul element a
flex container and will allow each of the items inside of it to
become flex items. This enables us to position these flex items as we
please in order to make a visually appealing and responsive site.