Edit class names, create flex container
The class names previously added were not added to the right elements
so they have been edited so that they target the right elements now.
The class selector for the "right" class has been added alongside the
display, align-items and gap properties in order to make the div
element a flex container that has vertically aligned flex items with
equal gaps between them.