Remove class names, add auto margin
The class names previously added to index.html have been removed. This
is due to the fact that these values were unnecessary so have been
removed.

The class selector "notifications" has been added to style.css in
order to add an auto margin to the left side of the button element.
This will allow the button and profile image div element to stay on
the right hand side of the screen no matter how the window is sized,
increasing the site's responsiveness.