Add div element surrounding media-container divs
A div element has been added around the media-container divs so that
each media-container div is nested within it. This div has been
assigned the class "information". This will allow us to make this div
a flex container to arrange the individual media-container divs within
it as they will then become flex items.