Create flex container
The display property has been assigned to the information class
selector with the value of flex. This means that the div with the
corresponding class attribute and value is now a flex container. This
will be useful as each div nested directly within the information div
will now be a flex item that we can position in any manner we would
like using flexbox.