Add flex shorthand to media-container divs
The flex shorthand property has been applied to the media-container
class selector. There is no growth factor so the items will not grow
past the flex basis of 200px, which ensures that the items remain at
a width of 200px and will not grow past this point. There is a shrink
factor of one so that on smaller screens, the text shrinks to fit the
container.