Add font-weight to buttons and header
The font-weight property with value set to bold has been added to both
the continue and cancel buttons.

The font-weight property with value set to bold has also been added to
the newly created header class selector in order to emphasise this
text. The font-size property has also been added to this class
selector in order to make the header stand out from the rest of the
text.