Add div, rearrange divs
A div element has been added around the continue and cancel buttons
and has been assigned the class attribute with a value of "buttons".
This will allow us to individually use flex to target these
buttons.

The div elements have also been rearranged so that each element
appears where it should on the page.