Create flex containers, add flex-flow value
Flex containers have been created to group parts of the page together
so that they can be appropriately targeted by flexbox properties and
values.

The flex-flow property has been added to the information class with a
value of "column wrap" in order to ensure that the text appears below
the header and wraps appropriately.