Remove gap, add margins and add flex shorthand
The gap property has been removed from the modal class as it was
deemed unnecessary.

Margins have been added: on the right for the icon class, on the
bottom for the header class and on the right and bottom for the text
class.

The flex shorthand property has also been added to the text class in
order to allow the text to grow/shrink to fill its available container
space.