Remove bullets, margin and padding from ul
The properties margin and padding have been added to the type selector
ul with a value of 0 in order to get rid of any automatic margin and
padding added to these elements.

The list-style property has also been added with a value of none to
remove the bullets before each list item.