Remove text decoration for a, change link colour
The text-decoration property has been added to the type selector a and
has been set to a value of none. This will remove the underlining of
all anchor elements.

The colour of these anchor elements have also been changed from blue
to a dark grey so that they fit the aesthetic of the website more.