Style content
Padding of 10px has been added to the buttons in order to make them
more noticeable and add space between the text and the border of the
content.

Flex containers have been created by adding the display: flex; rule to
the "content", "buttons" and "input" class selectors. This will allow
us to properly position them on the page and increase responsiveness.

The flex-flow of the content class has been changed to column wrap
while the input and buttons classes have kept the flex-flow of row
wrap. This allows each of the elements to stack above one another.

The image, input box and buttons have been centred using align-items
and justify-content properties with values set to center. This allows
the content to stay in the centre of the screen no matter the
dimensions.

A gap of 10px has also been added between the buttons in order to
ensure that they are spaced apart so that users see them clearly as
two seperate buttons.