Style footer
The body of the page has been turned into a flex container in order to
stick the footer to the bottom of the page. The flex direction of the
body has also been changed to column to ensure the footer is stacked
below all other content.

The margin of the content class has been changed from 50px. There are
now two properties: margin-top and margin-bottom, with values both set
to auto to ensure the content stays in the centre of the page.

The footer flex container has been created and a margin-top property
with the value set to auto has been added to this container. This
ensures that the footer will be stuck to the bottom of the page no
matter what screen dimensions are being used.

The footer has been styled using padding and background colour. A
padding value of 10px has been added to ensure the background colour
fills the entirety of the footer and that there is space between the
edge of the page and the links. A light grey background colour has
been added to make the footer stand out from the rest of the page.