Remove margin from header, add gap
The margin property has been removed from the header and replaced with
a padding of 10px in order to keep the styling between the header and
footer consistent.

A gap of 10px has also been added between the image, input and buttons
in the content class. This ensures everything is spaced out nicely and
is easily readable and accesssible.