Remove bullets and default margin/padding
The list-style-type property has been added to the type selector ul
with the value none in order to remove the bullets before each list
item.

The default margin and padding have been removed by setting the margin
property to 0px and the padding property for ul and li to 5px each.
This allows us to define the space between each list item and the edge
of the ul element.