Create flex containers
Three flex containers have been created for the "sidebar", "content"
and "cards" classes. This will allow us to position the flex items
they contain in a way that is similar to the assignment's requirements.

The flex direction of the sidebar flex container has been changed to
column so that each item is displayed below one another.

The flex flow of the cards flex container has been changed to row wrap
to ensure that each card displays next to one another and wraps onto
the next flex line if needed. In order to ensure that flex wrap works
properly, the cards have each been assigned a width of 300px.