Create flex container and style header
The flex container for the header class has been created. This will
allow us to position the header using flexbox as required.

The text contained in the header divs has been vertically aligned
using flex. The text has also been appropriately styled using a font
size of 30px and a font weight of 900 in order to make it stand out
from the rest of the text on the page.

The box-sizing property has been set to border-box so that we are able
to work using the alternate box model. This will make it easier to
check the height of the header div element.

Padding has been added to the header div element in order to ensure
that there is space between the text and the edge of the container.