Stick footer to bottom of screen and align text
The body element has been made into a flex container with the flex
direction set to column. This is to make sure that the footer is able
to stick to the bottom of the screen despite any dimension changes.

The margin-top property has been added to the footer class selector
with a value of auto. This ensures the footer is able to stretch to
the bottom of the screen at all times.

The footer class has also been made into a flex container. This has
been done so that it can be positioned as required. The text has then
been centrally aligned using both the justify-content and align-items
properties.