Remove auto margin, add flex shorthand
The margin-top: auto; rule has been removed from the footer class
selector as it is unnecessary and was creating unwanted space between
the content and the footer.

In order to stick the footer to the bottom without this auto margin,
the flex: 1; shorthand rule has been added to the content class
selector in order to ensure it takes up all available space. This
means that no matter the dimensions, the content will fill the main
body of the page and the footer will appear below it, stuck to the
bottom of the page.